Roshan_Internship/Roshan_Task03_01_00/Task03_01_00_News_Builder_API/requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.5
cssselect==1.3.0
defusedxml==0.7.1
Django==5.2.4
django-celery-beat==2.8.1
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
exceptiongroup==1.3.0
filelock==3.18.0
flower==2.0.1
gunicorn==23.0.0
h11==0.16.0
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.0
itemloaders==1.3.2
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.0
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
queuelib==1.8.0
redis==6.2.0
requests==2.32.4
requests-file==2.1.0
Scrapy==2.13.3
selenium==4.34.2
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tldextract==5.3.0
tomli==2.2.1
tornado==6.5.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
